goals list added

+++ features/goals/src/main/java/com/ikcollab/goals/components/GoalItem.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -22,7 +21,7 @@ fun GoalItem(
     end:Long,
     stepsCount:Int,
     stepsCompletedCount:Int,
-    daysLeft:Int,
+    daysLeft: Long,
 ) {
     Box(modifier = modifier
         .fillMaxWidth()
